move get notif to root
